Logic to show the current receipt and update the receipt

+++ app/src/main/java/com/recippie/doctor/app/data/ReceiptData.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "receiptdata")
 data class ReceiptData(
     @ColumnInfo(name = "num_medicine") @PrimaryKey (autoGenerate=true) var numMedicine: Int = 0,
-    @ColumnInfo(name = "num_receipt") val numReceipt: Int,
+    @ColumnInfo(name = "num_receipt") val numReceipt: Long,
     @ColumnInfo(name = "medicine") val description: String,
     @ColumnInfo(name = "each_time") val eachTime: String,
     @ColumnInfo(name = "during_time") val duringTime: String
